Add detection of parameterized Puppet classes
This change expands one of the regular expressions used to disambiguate
Puppet and Pascal files ending with ".pp". Without this, parameterized
Puppet classes do not match and end up being detected as Pascal.

I have confirmed that this regular expression matches everything in my
Puppet tree that it matched with the previous, and all that failed to
match but should have, save one -- a file with nothing but comments.